### 2023-03-01 #5168 @crvernon
Great @jlevente! We aim to finish up the full review process in 4-6 weeks.  This includes author responses to reviewer comments as well.  This whole part is (thankfully!) different than other journals that you may have encountered though because it is done iteratively here on GitHub.  So I think you will find the initial review feedback part very straightforward and quick.
