### 2023-03-01 #5201 @editorialbot
[REVIEW]: Mobilkit: A Python Toolkit for Urban Resilience and Disaster Risk Management Analytics using High Frequency Human Mobility Data
**Submitting author:** @takayabe0505 (<a href="http://orcid.org/0000-0001-8967-1967">Takahiro Yabe</a>)
**Repository:** https://github.com/mindearth/mobilkit
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.8
