### 2023-04-27 #5201 @crvernon
🚨 MIDWEEK RALLY 🚨

@ssujit and @jlevente - Could you please provide an update to how things are going here?  Please do let me know if you are unable to review this article so I may move it forward. As always, let me know if you have any questions!

Have a great day!
