### 2023-06-04 #5201 @crvernon
👋 @ifthompson and  @levisweetbreu

Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the [target repository](https://github.com/mindearth/mobilkit) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

**Reviewer instructions & questions**
@ifthompson  & @levisweetbreu , your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.

`@editorialbot generate my checklist`

The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let me know.
