### 2023-06-09 mindearth/mobilkit#4 @ubi15
Hi @levisweetbreu, at the time being there are no automatic tests in the code. The reference for the user is in the tutorial notebook [mobilkit_tutorial.ipynb](https://github.com/mindearth/mobilkit/blob/main/docs/examples/mobilkit_tutorial.ipynb) to check on an open dataset that the code is running as expected.

Please let me know if you need other info.
