### 2023-06-12 mindearth/mobilkit#5 @levisweetbreu
Author Contributions
@takayabe0505 or @ubi15 could you give me a brief description of each author's contribution to the software/paper?

https://github.com/openjournals/joss-reviews/issues/5201
