### 2023-06-14 mindearth/mobilkit#4 @levisweetbreu
@ubi15 I'll accept the manual steps you have shown as it falls under the 'OK' section of the [JOSS review guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). However, I strongly recommend adding automated tests so that the user can verify each function in `mobilkit` works as intended especially when new features are added or the source code is modified. See [this repository](https://github.com/adrn/gala/tree/main) which has a good example of an automated test suite.
