### 2023-06-14 mindearth/mobilkit#6 @levisweetbreu
Mobilkit Tutorial
@takayabe0505 and @ubi15 I have reviewed [mobilkit_tutorial.ipynb](https://github.com/mindearth/mobilkit/blob/main/docs/examples/mobilkit_tutorial.ipynb) to assess the functionality of the package and I have a few thoughts:

1. When merging the synthetic data with the empirical data, the `merge` function does not work as shown because the `uid` column is acting as the dataframe's index by default, and the `nunique` function does not work on the index column. I was able to bypass this by resetting the index before using `groupby` and `agg`, and the head of the dataframe changed slightly but the following plot did not change meaningfully.
2. Also when merging the synthetic data with the empirical the documentation states that it is data for Mexico, but it should be for Beijing, correct?
3. Is the Mexican data used for the remainder of the tutorial available for download? It looks like the CSV used is produced in one of the example notebooks ([mobilkit_population_per_area.ipynb](https://github.com/mindearth/mobilkit/blob/main/examples/02_mobilkit_population_per_area.ipynb)) but when I follow the trail I cannot find the original input data ("/data/dataHFLB/mexico/20171*/part-*.csv.gz"). If the input data is available, the tutorial should tell users how to create the CSV, and if the input data is not available for download, the tutorial should state that.

https://github.com/openjournals/joss-reviews/issues/5201
