+++ setup.cfg
@@ -30,8 +30,8 @@ classifiers =
 packages = mobilkit
 python_requires = >=3.6
 install_requires =
-	distributed[full]==2022.4.1
-	dask[full]==2022.4.1
+	distributed[full]==2023.5.0
+	dask[full]==2023.5.0
 	click
 	geopandas
 	descartes
