### 2023-06-20 mindearth/mobilkit#5 @takayabe0505
Sure, here it is:
Enrico Ubaldi -- code development, data analysis, visualization, writing
Takahiro Yabe -- code development, data analysis, visualization, writing
Nicholas Jones -- data analysis, review of writing, supervision
Maham Faisal Khan -- data analysis, review of writing
Alessandra Feliciotti -- data analysis, review of writing
Riccardo Di Clemente -- data analysis, review of writing
Satish V. Ukkusuri -- data analysis, review of writing, supervision
Emanuele Strano -- data analysis, review of writing, supervision
