### 2023-06-27 mindearth/mobilkit#8 @ifthompson
License Typo?
Hi @takayabe0505, I am going through the JOSS review checklist and noticed that the LICENSE file says "MinEarth", rather than "MindEarth". Is this a typo?

https://github.com/openjournals/joss-reviews/issues/5201
