### 2023-07-03 mindearth/mobilkit#9 @ifthompson
Tutorial Beijing Shapefile
In this section of the tutorial, https://mobilkit.readthedocs.io/en/latest/examples/mobilkit_tutorial.html#Create-tessellation-from-points, the file "DT19new/PopCensus2010_township.shp" is not in the data downloaded from the link, which appears to be for a different dataset "DT41". I was able to find the files for DT19 (https://www.beijingcitylab.com/app/download/12184598199/DT19_v2.zip?t=1644650253) on the website, and then follow the tutorial as expected, so perhaps the link in the tutorial needs to be changed.

https://github.com/openjournals/joss-reviews/issues/5201
