### 2023-07-18 mindearth/mobilkit#9 @ubi15
Hi @ifthompson  , thanks for spotting this typo, we fixed it in ab1c93c.

Closing this issue.
