### 2023-07-25 #5201 @ifthompson
@crvernon aside from what @levisweetbreu has mentioned, I think the community guidelines just need to explicitly mention that users are welcome to raise an issue to `2) Report issues or problems with the software 3) Seek support`, as per the review checklist, not just to contribute. @takayabe0505 let me know if I have missed this somewhere in the documentation, otherwise it should be a quick add to the README.

Not a blocking issue, but non-specialists reading the paper might benefit from the term "mobility data" being defined early on. I think from context it's clear enough to understand, but for those outside of this field, "mobility" can be more associated with as the ability to physically move (e.g., move your body's joints), rather than high frequency location data.

@crvernon One last thing: JOSS's Human participants research policy primarily relates to medical data, rather than privacy/ethics concerns with this kind of location data, such as determining home/work location. The documentation mentions that mobility data often has noise added near the user's home in order to preserve their privacy, so home location cannot be determined with higher resolution than this noise if present, but does JOSS have any guidelines relating to software that works with this kind of data?

Once these are addressed I think this is good.
