### 2023-08-12 #5201 @takayabe0505
Also, we have addressed the following comment

> One comment is: I think the community guidelines just need to explicitly mention that users are welcome to raise an issue to 2) Report issues or problems with the software 3) Seek support, as per the review checklist, not just to contribute. @takayabe0505 let me know if I have missed this somewhere in the documentation, otherwise it should be a quick add to the README.

Please see the [7b50b0e](https://github.com/mindearth/mobilkit/commit/7b50b0e47b648e41c28e25602540a9aebe6d028c) commit, thanks!
