### 2023-08-23 92c5aef
Update .readthedocs.yaml to stick to build.os.
