+++ .readthedocs.yaml
@@ -4,7 +4,11 @@
 
 # Required
 version: 2
-
+build:
+  os: "ubuntu-22.04"
+  tools:
+    python: "3.8"
+    
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    builder: html
