+++ .readthedocs.yaml
@@ -7,7 +7,7 @@ version: 2
 build:
   os: "ubuntu-22.04"
   tools:
-    python: "3.8"
+    python: "3.11"
     
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
@@ -21,6 +21,6 @@ formats:
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-   version: 3.8
+   version: 3.11
    install:
    - requirements: docs/requirements.txt
