### 2023-08-23 #5201 @levisweetbreu
@crvernon I need input from you on the issues I mentioned in my previous comment, namely the authors' use of proprietary data in the paper and for verification of the software.
