+++ setup.cfg
@@ -1,7 +1,7 @@
 
 [metadata]
 name = mobilkit
-version = 0.2.6
+version = 0.2.7
 author = MindEarth
 author_email = enrico.ubaldi@mindearth.org
 description = A Python Toolkit for Urban Resilience and Disaster Risk Management Analytics using High Frequency Human Mobility Data
