### 2023-11-08 #5201 @takayabe0505
Hi @crvernon sorry for the late response. We used open source data (there's only 1 publicly available open dataset we know of - the GeoLife data; https://www.microsoft.com/en-us/research/publication/geolife-gps-trajectory-dataset-user-guide/) to make sure our functions work properly, in the 'Quickstart' section in the documentation:
https://mobilkit.readthedocs.io/en/latest/examples/mobilkit_tutorial.html#

I hope this satisfies the reviewers' concerns. Thanks!
