### 2023-11-10 #5201 @crvernon
@takayabe0505 there are still two open issues related to this thread:

https://github.com/mindearth/mobilkit/issues/5

https://github.com/mindearth/mobilkit/issues/6

Have these been satisfied?
