### 2023-12-01 #5201 @crvernon
Some final requests to your paper on per my last re-read:

- [ ] LINES 18-20.  Should be rephrased, "While several tookits allow users to perform some basic analytics on large mobility datasets (e.g., De Montjoye et al., 2016; Pappalardo et al., 2019), these cover only some of the steps in the mobility data pipeline."
- [ ] LINES 21-22.  Rephrase the following "Very often users have to go fishing for functions/tools from other libraries to cover for data pre-processing or visualization." to "These toolkits also do not provide adequate data pre-processing and visualization functionality which causes users to seek additional external options."
- [ ] LINE 23.  Change "...documentation that enables..." to "...documentation to enable..."
- [ ] LINE 34.  Change "...Notebooks that provides..." to "...notebooks that provide..."
- [ ] LINES 40-48.  Create a new section titled "Use Case" to hold this content.  Here is a revised version of these lines that reads more clearly:  "The usefulness of `Mobilkit` was demonstrated in a recent study carried out in collaboration with the World Bank Global Facility for Disaster Reduction and Recovery.  The study focused on assessing the impact of a 7.1 magnitude earthquake that occurred on September 19, 2017 where the epicenter was located around 55 km south of Puebla, Mexico (about 100 km south-east of Mexico City, Mexico).  `Mobilkit` was used to conduct an analysis of the spatial structure of ten surrounding cities from smartphone location data, provided by Quadrant, which was collected before and after the earthquake to generate insights about management options.  Similar analysis could also be explored using `Mobilkit` for planning and recovering activities related to climate, man-made, and other natural disasters."
**Please review my revisions in the above paragraph for accuracy**

Once you have made these changes we can move forward.  You do not have to conduct an additional code release or create any new DOI.  Just alter the paper on your current, up-to-date branch that is being used in this review.

Thanks!
