### 2023-12-05 014fde2
Update paper.md to reflect reviewers feedbacks.
