### 2023-12-06 #5201 @ubi15
Dear Chris,
Thanks a lot for your review, we inserted all the changes in the paper as requested.

Below, you will find a report of the edits we performed as well as of the amendments applied to your suggestions.

We already pushed the commits to the main branch (commits [96761c](https://github.com/mindearth/mobilkit/commit/014fde21267a939a1092e2812e94dd122896761c) and [936eca](https://github.com/mindearth/mobilkit/commit/d965ebaacaf5ad7e598cb1ce6c2949b966936eca)).

Feel free to write to us should you need any other edits or information.

Thanks!
Enrico on behalf of the team

--------
### Edits

- [x] LINES 18-20. No edits, inserted as requested.
- [x] LINES 21-22. No edits, inserted as requested.
- [x] LINE 23. No edits, inserted as requested.
- [x] LINE 34. No edits, inserted as requested.
- [x] LINES 40-48. Some edits: we created the new section titled "Use Case" and changed the content to reflect the fact that we are referring to **two separate works**: one concerning the **Puebla earthquake study** -that we also added to the references- and the other dealing with **Urban Spatial Structure** -covering ten cities around the world.  We added a footnote pointing readers to the Notebooks covering the latter topic as well. We are reporting here the revised version of the lines you suggested:

> The usefulness of `Mobilkit` was demonstrated in a recent study carried out in collaboration with the World Bank Global Facility for Disaster Reduction and Recovery[@yabe2021location]. The study focused on assessing the impact of a 7.1 magnitude earthquake that occurred on September 19, 2017 where the epicenter was located around 55 km south of Puebla, Mexico (about 100 km south-east of Mexico City, Mexico). `Mobilkit` was also leveraged to conduct an analysis of the spatial structure of ten cities around the globe using smartphone location data, provided by Quadrant, to generate insights about mobility management options[^1]. Similar analysis could also be explored using `Mobilkit` for planning and recovering activities related to climate, man-made, and other natural disasters.
> [^1] See the notebooks covering [Urban Spatial Structure analyses](https://mobilkit.readthedocs.io/en/latest/examples/USS01_Mumbai.html) and an [inter-city comparison of Urban Spatial Structure indicators](https://mobilkit.readthedocs.io/en/latest/examples/USS02_CityComparison.html).

And here is the bibtex of the added reference:
```bibtex
@misc{yabe2021location,
      title={Location Data Reveals Disproportionate Disaster Impact Amongst the Poor: A Case Study of the 2017 Puebla Earthquake Using Mobilkit},
      author={Takahiro Yabe and Nicholas K W Jones and Nancy Lozano-Gracia and Maham Faisal Khan and Satish V. Ukkusuri and Samuel Fraiberger and Aleister Montfort},
      year={2021},
      eprint={2107.13590},
      archivePrefix={arXiv},
      primaryClass={physics.soc-ph}
}
```
