### 2024-01-14 #5201 @crvernon
Thanks for making the changes @ubi15 though we still have one more minor thing to fix in the paper before moving on to the next steps...

- [ ] LINE 47:  You are generating a footnote but it is not rendering correctly.  Please ensure that 1) in the text that there is not space between `options` and `[^1]` which I believe you have accounted for; and 2) the footnote itself has the colon after it (e.g., `[^1]: See the notebook...`).

Here is an generic example:

```
Articles are published under a Creative Commons license[^1].
Software should use an OSI-approved license.

[^1]: An open license that allows reuse.
```

Thanks and let me know when you have this fixed up!
