### 2024-02-10 #5201 @oliviaguest
@takayabe0505 can you fix the double parentheses for the citations please? So make it for example:
> (e.g., De Montjoye et al., 2016 or Pappalardo et al., 2019)

Otherwise, it's looking good. Just tag me when you're done. :)
