### 2024-02-15 #5201 @ubi15
@oliviaguest we substituted the parentheses with an em dash (as the inner parenthesis is due to the citation style).

Let us know if you need any other edit.
